rename module and go fmt
